fix by components
